backend/requirements.txt
channels
channels_redis
daphne
django
django-restframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
# psycopg2
psycopg2-binary
pillow
redis